src/requirements.txt
google-genai
dotenv
flask
fastapi[standard-no-fastapi-cloud-cli]